split  to files
